working through iOS splits
