+++ Platforms/iOS/PlatformBluetooth.cs
@@ -1,11 +1,24 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using EnlightenMAUI.Models;
 
 namespace EnlightenMAUI.Platforms.iOS
 {
     internal class PlatformBluetooth
     {
-        PendingIntent usbIntent;
+        internal static async Task<List<USBViewDevice>> doUSBScanAsync()
+        {
+            Logger logger = Logger.getInstance();
+            List<USBViewDevice> usbDeviceList = new List<USBViewDevice>();
+
+            return usbDeviceList;
+        }
+
+        internal static async Task<bool> doConnectOrDisconnectUSBAsync(Spectrometer spec)
+        {
+            return false;
+        }
+
     }
 }
