+++ ViewModels/BluetoothViewModel.cs
@@ -14,6 +14,14 @@ using EnlightenMAUI.Common;
 using LibUsbDotNet.Main;
 using Microsoft.Maui;
 
+#if ANDROID
+using EnlightenMAUI.Platforms.Android;
+#endif
+
+#if IOS
+using EnlightenMAUI.Platforms.iOS;
+#endif
+
 namespace EnlightenMAUI.ViewModels;
 
 public class BluetoothViewModel : INotifyPropertyChanged
@@ -420,55 +428,10 @@ public class BluetoothViewModel : INotifyPropertyChanged
     private async Task<bool> doUSBScanAsync()
     {
         logger.info("Looking for usb devices via Android services");
-        try
-        {
-            Context con = Android.App.Application.Context;
-            UsbManager manager = (UsbManager)con.GetSystemService(Context.UsbService);
+        List<USBViewDevice> temp = await PlatformBluetooth.doUSBScanAsync();
+        usbDeviceList = new ObservableCollection<USBViewDevice>(temp);
 
-            var features = con.PackageManager.GetSystemAvailableFeatures();
-            foreach ( var feature in features ) 
-            {
-                logger.info("{0} feature available", feature.Name);
-            }
-
-            if (manager.DeviceList.Count == 0)
-            {
-                logger.info("No USB devices found");
-            }
-
-            foreach (Android.Hardware.Usb.UsbDevice acc in manager.DeviceList.Values)
-            {
-                this.acc = acc;
-
-                String desc = String.Format("Vid:0x{0:x4} Pid:0x{1:x4} (rev:{2}) - {3}",
-                    acc.VendorId,
-                    acc.ProductId,
-                    acc.Version,
-                    acc.DeviceName);
-
-                logger.info("found usb device {0}", desc);
-
-                if (acc.VendorId == 0x24aa)
-                {
-                    USBViewDevice uvd = new USBViewDevice(acc.DeviceName, acc.VendorId.ToString("x4"), acc.ProductId.ToString("x4"));
-                    usbDeviceList.Add(uvd);
-
-                    usbIntent = PendingIntent.GetBroadcast(con, 0, new Intent(ACTION_USB_PERMISSION), PendingIntentFlags.Immutable);
-                     
-                    //LibUsbDotNet.UsbDevice usbDevice = LibUsbDotNet.UsbDevice.OpenUsbDevice(d => d.Pid == acc.ProductId);
-                    manager.RequestPermission(acc, usbIntent);
-                    
-                }
-
-            }
-        }
-        catch (Exception ex)
-        {
-            logger.info("USB grab failed with error {0}", ex.Message);
-        }
-
-        return true;
-        
+        return true;        
     }
 
     async Task<bool> _requestPermissionsAsync()
@@ -600,90 +563,18 @@ public class BluetoothViewModel : INotifyPropertyChanged
     
     private async Task<bool> doConnectOrDisconnectUSBAsync()
     {
-        if (spec == null || (spec is BluetoothSpectrometer))
+        bool ok = await PlatformBluetooth.doConnectOrDisconnectUSBAsync(spec);
+
+        if (ok)
         {
-            try
+            if (spec != null && spec.paired)
             {
-                buttonConnectEnabled = false;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(connectButtonBackgroundColor)));
-                Context con = Android.App.Application.Context;
-                UsbManager manager = (UsbManager)con.GetSystemService(Context.UsbService);
-                connectionProgress = 0.025;
-                int interfaces = acc.InterfaceCount;
-                connectionProgress = 0.05;
-
-                logger.info("usb device has {0} interfaces", interfaces);
-                for (int i = 0; i < interfaces; i++)
-                {
-                    logger.info("interface {0} has {1} endpoints", i, acc.GetInterface(i).EndpointCount);
-                }
-                connectionProgress = 0.075;
-
-                UsbDeviceConnection udc = manager.OpenDevice(acc);
-                connectionProgress = 0.1;
-                logger.info("device has {0} configurations", acc.ConfigurationCount);
-                if (udc != null)
-                {
-                    logger.info("successfully opened device");
-                    USBSpectrometer usbSpectrometer = new USBSpectrometer(udc, acc);
-                    connectionProgress = 0.125;
-                    spec = usbSpectrometer;
-                    spec.showConnectionProgress += showSpectrometerConnectionProgress;
-
-                    connectionProgress = 0.15;
-                    bool ok = await (spec as USBSpectrometer).initAsync();
-                    if (ok)
-                    {
-                        logger.debug("invoking new connection");
-                        if (Spectrometer.NewConnection != null)
-                            Spectrometer.NewConnection.Invoke(this, spec);
-                    }
-                    logger.debug("init complete setting instance and paired");
-                    USBSpectrometer.setInstance(usbSpectrometer);
-                    USBViewDevice.paired = true;
-                    connectionProgress = 1;
-                    buttonConnectEnabled = true;
-                    PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(connectButtonBackgroundColor)));
-                    Settings.getInstance().spec = spec;
-                    await Shell.Current.GoToAsync("//ScopePage");
-                    connectionProgress = 0;
-                    return ok;
-                }
-                else
-                {
-                    logger.info("failed to open device");
-                    return false;
-                }
-            }
-            catch (Exception ex)
-            {
-                logger.error("USB connect failed with exception {0}", ex.Message);
-                return false;
-            }
-
-        }
-        else
-        {
-            logger.info("already initialized as usb spec, reconnecting");
-
-            try
-            {
-                if ((spec as USBSpectrometer).paired)
-                    spec.disconnect();
-                else
-                {
-                    (spec as USBSpectrometer).connect();
+                if (Spectrometer.NewConnection != null)
                     Spectrometer.NewConnection.Invoke(this, spec);
-                    return await (spec as USBSpectrometer).initAsync();
-                }
-                return true;
-            }
-            catch (Exception ex)
-            {
-                logger.error("USB toggle failed with exception {0}", ex.Message);
-                return false;
             }
         }
+
+        return ok;
     }
 
     async Task<bool> doDisconnectAsync(bool isBluetooth)
